cr_data_preprocessing/requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
black==24.2.0
blinker==1.7.0
boltons==24.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cobble==0.1.3
datasets==2.18.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
environs==9.5.0
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.2.0
Genshi==0.7.7
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
htmldiff==0.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
Mako==1.3.2
mammoth==1.7.1
MarkupSafe==2.1.5
marshmallow==3.21.0
minio==7.2.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
peft==0.10.0
pillow==10.3.0
platformdirs==4.2.0
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.2
pydantic_core==2.16.3
pymilvus==2.3.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.1
sentry-sdk==1.44.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
wandb==0.16.5
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
